+++ FSIC/__init__.py
@@ -11,3 +11,5 @@ models in the tradition of Wynne Godley.
 
 from FSIC.version import MAJOR, MINOR, PATCH, DEV, VERSION as __version__  # noqa
 from FSIC.settings import DTYPE  # noqa
+
+from FSIC.classes.schematic import Variable